% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creat_Project.R
\name{creat_project}
\alias{creat_project}
\title{Title Creat Project}
\usage{
creat_project(ProjectName)
}
\arguments{
\item{ProjectName}{Your project name.}
}
\description{
Title Creat Project
}
\examples{
creat_project(ProjectName = "ATAC")
}
