% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_limma.R
\name{deg_limma}
\alias{deg_limma}
\title{Title deg_limma}
\usage{
deg_limma(
  expr,
  group_info,
  contrast1 = "tumor",
  contrast2 = "normal",
  p = 0.05,
  logfc = 1
)
}
\arguments{
\item{logfc}{}
}
\description{
Title deg_limma
}
