% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_xena.R
\name{download_xena}
\alias{download_xena}
\title{Title download_xena}
\usage{
download_xena(project, data_type = NULL)
}
\arguments{
\item{data_type}{}
}
\value{
a Rdata file with clinical, survival, and mRNA expression.
}
\description{
Title download_xena
}
\examples{
download_xena(project = 'LUAD',data_type = 'counts')
}
