% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_xena_maf.R
\name{download_Xena_maf}
\alias{download_Xena_maf}
\title{Title download_Xena_maf}
\usage{
download_Xena_maf(project = "eg:STAD")
}
\arguments{
\item{project}{}
}
\value{
A dataframe for MuTect2 data frome xena, which can be deal with maftools.
}
\description{
Title download_Xena_maf
}
\examples{
download_maf(project = 'LUAD')
}
