% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_gene_cor.R
\name{One_gene_cor}
\alias{One_gene_cor}
\title{One_gene_cor}
\usage{
One_gene_cor(expr, gene, p.threshold = 0.05, parallel = FALSE)
}
\arguments{
\item{p.threshold}{}
}
\value{
a list with all correlation and significant correlation
}
\description{
To calculate correlation with others for one gene.
}
\examples{
cor_res = One_gene_cor(expr = expr,gene = gene,p.threshold = 0.05)
}
