% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_gene_deg.R
\name{One_gene_deg}
\alias{One_gene_deg}
\title{title One_gene_deg}
\usage{
One_gene_deg(expr, gene, p = 0.05, logfc = 1)
}
\arguments{
\item{logfc}{}
}
\value{
A list with deg_list and deg_sig.
}
\description{
To caculate the differtial expressed gene by limma.
}
\examples{
res = diffExprAnalysis(expr = expr, gene = gene,p = 0.01, logfc = 0.5)
}
