% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_gene_ssgsea.R
\name{One_gene_ssgsea}
\alias{One_gene_ssgsea}
\title{Title One_gene_ssgsea}
\usage{
One_gene_ssgsea(gene, expr, geneset)
}
\arguments{
\item{geneset}{}
}
\description{
Title One_gene_ssgsea
}
\examples{
gsea_res <- run_ssgsea(expr = expr.tpm,gene=gene,geneset = gene_set)
}
