% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Title plot_bar}
\usage{
plot_bar(
  data,
  Group1 = NULL,
  Group2 = NULL,
  values = NULL,
  compare_list = NULL,
  signif.method = "t.test",
  legend_position = "right",
  plot_line = F,
  plot.title = NULL,
  x.title = NULL,
  y.title = NULL,
  signif.position.y,
  pal = NULL,
  map_signif = T
)
}
\arguments{
\item{data}{A dataframe which include at least one group and the corresponding value}

\item{Group1}{The first group you focuse on}

\item{Group2}{The second group you focuse on}

\item{values}{The values of your data}

\item{compare_list}{The group to compare, the default is to compare all group}

\item{signif.method}{The statistical method to use, default is 't.test', also 'wilcox.test'}

\item{signif.position.y}{The significant mark position}

\item{pal}{The color matching plate, you can provide a color vector to change the color}
}
\description{
Title plot_bar
}
\examples{
plot_bar(data,Group1  = "group1" ,Group2 = 'group2', values  =  "value")

}
