% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxblot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Title plot_boxblot}
\usage{
plot_boxplot(
  dataset,
  group,
  values,
  title = "",
  x.title = "",
  y.title = "",
  color = c("#5CB85C", "#337AB7", "#F0AD4E", "#D9534F")
)
}
\arguments{
\item{color}{}
}
\description{
Title plot_boxblot
}
\examples{
plot_boxplot(dataset, group, values, title = "Boxplot of Sepal Length", x.title = "Species", y.title = "Sepal Length")
}
