% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deg_volcano.R
\name{plot_deg_volcano}
\alias{plot_deg_volcano}
\title{Title plot_deg_volcano}
\usage{
plot_deg_volcano(
  plot_data,
  logFC_name,
  pvalue_name,
  p.cutoff = 0.05,
  label_text_size = 3,
  custum_label = NULL
)
}
\arguments{
\item{custum_label}{}
}
\description{
Title plot_deg_volcano
}
\examples{
plot_deg_volcano(plot_data = deg_res,logFC_name = 'logFC',pvalue_name = "P.Value",label_text_size = 4,custum_label = c('Cd86'),p.cutoff = 0.05)
}
