% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scRNA_dotplot.R
\name{plot_scRNA_dotplot}
\alias{plot_scRNA_dotplot}
\title{Title plot_scRNA_dotplot}
\usage{
plot_scRNA_dotplot(
  scRNA,
  features,
  legend.position = "right",
  plot.title = "",
  title.x = "",
  title.y = ""
)
}
\arguments{
\item{scRNA}{A seurat object}

\item{features}{A gene vector to plot}

\item{legend.position}{right left}

\item{title.y}{}
}
\description{
Title plot_scRNA_dotplot
}
\examples{
plot_scRNA_dotplot(scRNA = scRNA_fibro,features = cell_marker)
}
