% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scRNA_featureplot.R
\name{plot_scRNA_featureplot}
\alias{plot_scRNA_featureplot}
\title{Title plot_scRNA_featureplot}
\usage{
plot_scRNA_featureplot(
  data,
  plot.gene = NULL,
  pal = NULL,
  pal.style = "D",
  X.title = NULL,
  y.title = NULL,
  pt.size = 1
)
}
\arguments{
\item{data}{A seurat object}

\item{plot.gene}{A vector including gene to plot}

\item{pal}{A vector with color, also you can use the default}

\item{pal.style}{The color style to plot, you can provide 'A' ... 'H'}

\item{X.title}{The title of axsis x}

\item{y.title}{The title of axsis y}

\item{pt.size}{The point size, the default is 1}
}
\description{
Title plot_scRNA_featureplot
}
\examples{
plot_scRNA_featureplot(data = seurat_obj,plot.gene = gene )
}
