% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ScRNA_reducion.R
\name{plot_ScRNA_reducion}
\alias{plot_ScRNA_reducion}
\title{Title plot_ScRNA_reducion}
\usage{
plot_ScRNA_reducion(
  scRNA,
  reduction = "umap",
  plot.column = "seurat_clusters",
  label = T,
  label.text.size = 6,
  cellcolors = c("#8B0000", "#E18727FF", "#0072B5FF", "#800080", "#BC3C29FF",
    "#20854EFF")
)
}
\arguments{
\item{scRNA}{A seurat object}

\item{reduction}{umap tsne}

\item{plot.column}{Which column in metadata to plot, default is "seurat_clusters"}

\item{label}{Whether to plot cell labels, default is TRUE}

\item{label.text.size}{The cell label size, default is 6}

\item{cellcolors}{A color vector. cellcolors=c("#8B0000","#E18727FF","#0072B5FF","#800080","#BC3C29FF","#20854EFF")}
}
\description{
Title plot_ScRNA_reducion
}
\examples{
plot_umap(scRNA = scRNA,plot.column = "group",label = F)
}
