% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scRNA_volcano.R
\name{plot_scRNA_volcano}
\alias{plot_scRNA_volcano}
\title{Title plot_scRNA_volcano}
\usage{
plot_scRNA_volcano(
  data,
  plot.gene = NULL,
  plot.gene.number = 5,
  label.text.size = 10,
  logfc.cut = 1,
  p.cut = 0.05
)
}
\arguments{
\item{data}{A seurat object}

\item{plot.gene}{A vector including gene to plot in the volcano plot}

\item{plot.gene.number}{The default will display the top 5 up and down regulated genes}

\item{label.text.size}{The size of the gene, default is 10}
}
\description{
Title plot_scRNA_volcano
}
