% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scRNA_monocle.R
\name{run_scRNA_monocle}
\alias{run_scRNA_monocle}
\title{Title run_scRNA_monocle}
\usage{
run_scRNA_monocle(scRNA, save_path = "./output_data/monocle2.Rdata")
}
\arguments{
\item{scRNA}{A Seurat Object}

\item{save_path}{A path to save your results}
}
\description{
Title run_scRNA_monocle
}
