% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xena_clean.R
\name{xena_clean}
\alias{xena_clean}
\title{title xena_clean}
\usage{
xena_clean(Rdata.path, project, data_type = "counts")
}
\arguments{
\item{data_type}{}
}
\description{
title xena_clean
}
\examples{
xena_clean(Rdata.path = "./output_xena/counts_xena_LUAD.Rdata",project = 'LUAD',data_type = 'counts')
}
