% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scRNA_volcano.R
\name{plot_scRNA_volcano}
\alias{plot_scRNA_volcano}
\title{Title plot_scRNA_volcano}
\usage{
plot_scRNA_volcano(
  data,
  plot.gene = NULL,
  plot.gene.number = 5,
  label.text.size = 80,
  logfc.cut = 1,
  p.cut = 0.05,
  label_postion_x_left = -2.5,
  label_postion_x_right = 2.5,
  label_postion_y = 0.95,
  base_size = 14,
  legend_size = 14,
  legend_position = "top"
)
}
\arguments{
\item{data}{A seurat object}

\item{plot.gene}{A vector including gene to plot in the volcano plot}

\item{plot.gene.number}{The default will display the top 5 up and down regulated genes}

\item{label.text.size}{The size of the gene, default is 10}
}
\description{
Title plot_scRNA_volcano
}
